期限日表示メソッド場所を変更
