+++ app/helpers/application_helper.rb
@@ -1,13 +1,2 @@
 module ApplicationHelper
-    require "date"
-
-    def diffDate(deadline_date)
-        today = Date.current
-        diff = (deadline_date - today).to_i
-        "あと#{diff}日"
-    end
-    
-    def today
-        today = Date.today
-    end
 end
